Column rule pack

Column rule pack is a rule pack which defines a set of rules for columns as a whole, i.e. functions which convert columns of interest to logical values. It should return a data frame with the following properties:

Number of rows equals to one.

Column names should be treated as concatenation of 'check column name' + 'separator' + 'rule name'.

Values indicate whether the column as a whole follows the rule.

Details

This format is inspired by dplyr's scoped variants of summarise() applied to non-grouped data.

The most common way to define column pack is by creating a functional sequence with no grouping and ending with one of:

summarise_all(.funs = rules(...)).

summarise_if(.predicate, .funs = rules(...)).

summarise_at(.vars, .funs = rules(...)).

Using rules()

Using rules() instead of funs() is recommended because:

It is a convenient way to ensure consistent naming of rules without manual name.

It adds a common prefix to all rule names. This helps in defining separator as prefix surrounded by any number of non-alphanumeric values.

See also

Data pack, group pack, row pack, cell pack.

Examples

# Validating present columns
numeric_column_rules <- . %>% dplyr::summarise_if(
  is.numeric,
  rules(mean(.) > 5, sd(.) < 10)
)
character_column_rules <- . %>% dplyr::summarise_if(
  is.character,
  rules(. %in% letters[1:4])
)

col_packs(
  num_col = numeric_column_rules,
  chr_col = character_column_rules
)

#> $num_col
#> A Column rule pack:
#> Functional sequence with the following components:
#> 
#>  1. dplyr::summarise_if(., is.numeric, rules(mean(.) > 5, sd(.) <     10))
#> 
#> Use 'functions' to extract the individual functions. 
#> 
#> $chr_col
#> A Column rule pack:
#> Functional sequence with the following components:
#> 
#>  1. dplyr::summarise_if(., is.character, rules(. %in% letters[1:4]))
#> 
#> Use 'functions' to extract the individual functions. 
#>