Create rule packs

Functions for creating different kinds of rule packs. Rule is a function which converts data unit of interest (data, group, column, row, cell) to logical value indicating whether this object satisfies certain condition. Rule pack is a function which combines several rules into one functional block. It takes a data frame of interest and returns a data frame with certain structure and column naming scheme. Types of packs differ in interpretation of their output.

data_packs(...)

group_packs(..., .group_vars, .group_sep = ".")

col_packs(...)

row_packs(...)

cell_packs(...)

Arguments

...

Functions which define packs. They can be in pure form or inside a list (at any depth).

.group_vars

Character vector of names of grouping variables.

.group_sep

String to be used as separator when uniting grouping levels for var column in exposure report.

Value

data_packs() returns a list of what should be data rule packs, group_packs() - group rule packs, col_packs() - column rule packs, row_packs() - row rule packs, cell_packs() - cell rule packs.

Details

These functions convert ... to list, apply rlang's squash() and add appropriate classes (group_packs() also adds necessary attributes). Also they are only constructors and do not check for validity of certain pack. Note that it is allowed for elements of ... to not have names: they will be computed during exposure. However it is a good idea to manually name packs.